Split renderer from camera.js
