+++ pose-detection/demos/live_video/src/camera.js
@@ -14,60 +14,19 @@
  * limitations under the License.
  * =============================================================================
  */
-import * as posedetection from '@tensorflow-models/pose-detection';
-import * as scatter from 'scatter-gl';
-
 import * as params from './params';
 import {isMobile} from './util';
 
-// These anchor points allow the pose pointcloud to resize according to its
-// position in the input.
-const ANCHOR_POINTS = [[0, 0, 0], [0, 1, 0], [-1, 0, 0], [-1, -1, 0]];
-
-// #ffffff - White
-// #800000 - Maroon
-// #469990 - Malachite
-// #e6194b - Crimson
-// #42d4f4 - Picton Blue
-// #fabed4 - Cupid
-// #aaffc3 - Mint Green
-// #9a6324 - Kumera
-// #000075 - Navy Blue
-// #f58231 - Jaffa
-// #4363d8 - Royal Blue
-// #ffd8b1 - Caramel
-// #dcbeff - Mauve
-// #808000 - Olive
-// #ffe119 - Candlelight
-// #911eb4 - Seance
-// #bfef45 - Inchworm
-// #f032e6 - Razzle Dazzle Rose
-// #3cb44b - Chateau Green
-// #a9a9a9 - Silver Chalice
-const COLOR_PALETTE = [
-  '#ffffff', '#800000', '#469990', '#e6194b', '#42d4f4', '#fabed4', '#aaffc3',
-  '#9a6324', '#000075', '#f58231', '#4363d8', '#ffd8b1', '#dcbeff', '#808000',
-  '#ffe119', '#911eb4', '#bfef45', '#f032e6', '#3cb44b', '#a9a9a9'
-];
 export class Camera {
-  constructor(useGpuRenderer) {
+  constructor() {
     this.video = document.getElementById('video');
-    this.canvas = document.getElementById('output');
-    this.ctx = useGpuRenderer ? null : this.canvas.getContext('2d');
-    this.scatterGLEl = document.querySelector('#scatter-gl-container');
-    this.scatterGL = new scatter.ScatterGL(this.scatterGLEl, {
-      'rotateOnStart': true,
-      'selectEnabled': false,
-      'styles': {polyline: {defaultOpacity: 1, deselectedOpacity: 1}}
-    });
-    this.scatterGLHasInitialized = false;
   }
 
   /**
    * Initiate a Camera instance and wait for the camera stream to be ready.
    * @param cameraParam From app `STATE.camera`.
    */
-  static async setupCamera(cameraParam, useGpuRenderer) {
+  static async setup(cameraParam) {
     if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
       throw new Error(
           'Browser API navigator.mediaDevices.getUserMedia not available');
@@ -92,7 +51,7 @@ export class Camera {
 
     const stream = await navigator.mediaDevices.getUserMedia(videoConfig);
 
-    const camera = new Camera(useGpuRenderer);
+    const camera = new Camera();
     camera.video.srcObject = stream;
 
     await new Promise((resolve) => {
@@ -109,166 +68,9 @@ export class Camera {
     camera.video.width = videoWidth;
     camera.video.height = videoHeight;
 
-    camera.canvas.width = videoWidth;
-    camera.canvas.height = videoHeight;
     const canvasContainer = document.querySelector('.canvas-wrapper');
     canvasContainer.style = `width: ${videoWidth}px; height: ${videoHeight}px`;
 
-    if (camera.ctx != null) {
-      // Because the image from camera is mirrored, need to flip horizontally.
-      camera.ctx.translate(camera.video.videoWidth, 0);
-      camera.ctx.scale(-1, 1);
-
-      camera.scatterGLEl.style =
-          `width: ${videoWidth}px; height: ${videoHeight}px;`;
-      camera.scatterGL.resize();
-
-      camera.scatterGLEl.style.display =
-          params.STATE.modelConfig.render3D ? 'inline-block' : 'none';
-    }
-
     return camera;
-  }
-
-  drawCtx() {
-    this.ctx.drawImage(
-        this.video, 0, 0, this.video.videoWidth, this.video.videoHeight);
-  }
-
-  clearCtx() {
-    this.ctx.clearRect(0, 0, this.video.videoWidth, this.video.videoHeight);
-  }
-
-  /**
-   * Draw the keypoints and skeleton on the video.
-   * @param poses A list of poses to render.
-   */
-  drawResults(poses) {
-    for (const pose of poses) {
-      this.drawResult(pose);
-    }
-  }
-
-  /**
-   * Draw the keypoints and skeleton on the video.
-   * @param pose A pose with keypoints to render.
-   */
-  drawResult(pose) {
-    if (pose.keypoints != null) {
-      this.drawKeypoints(pose.keypoints);
-      this.drawSkeleton(pose.keypoints, pose.id);
-    }
-    if (pose.keypoints3D != null && params.STATE.modelConfig.render3D) {
-      this.drawKeypoints3D(pose.keypoints3D);
-    }
-  }
-
-  /**
-   * Draw the keypoints on the video.
-   * @param keypoints A list of keypoints.
-   */
-  drawKeypoints(keypoints) {
-    const keypointInd =
-        posedetection.util.getKeypointIndexBySide(params.STATE.model);
-    this.ctx.fillStyle = 'Red';
-    this.ctx.strokeStyle = 'White';
-    this.ctx.lineWidth = params.DEFAULT_LINE_WIDTH;
-
-    for (const i of keypointInd.middle) {
-      this.drawKeypoint(keypoints[i]);
-    }
-
-    this.ctx.fillStyle = 'Green';
-    for (const i of keypointInd.left) {
-      this.drawKeypoint(keypoints[i]);
-    }
-
-    this.ctx.fillStyle = 'Orange';
-    for (const i of keypointInd.right) {
-      this.drawKeypoint(keypoints[i]);
-    }
-  }
-
-  drawKeypoint(keypoint) {
-    // If score is null, just show the keypoint.
-    const score = keypoint.score != null ? keypoint.score : 1;
-    const scoreThreshold = params.STATE.modelConfig.scoreThreshold || 0;
-
-    if (score >= scoreThreshold) {
-      const circle = new Path2D();
-      circle.arc(keypoint.x, keypoint.y, params.DEFAULT_RADIUS, 0, 2 * Math.PI);
-      this.ctx.fill(circle);
-      this.ctx.stroke(circle);
-    }
-  }
-
-  /**
-   * Draw the skeleton of a body on the video.
-   * @param keypoints A list of keypoints.
-   */
-  drawSkeleton(keypoints, poseId) {
-    // Each poseId is mapped to a color in the color palette.
-    const color = params.STATE.modelConfig.enableTracking && poseId != null ?
-        COLOR_PALETTE[poseId % 20] :
-        'White';
-    this.ctx.fillStyle = color;
-    this.ctx.strokeStyle = color;
-    this.ctx.lineWidth = params.DEFAULT_LINE_WIDTH;
-
-    posedetection.util.getAdjacentPairs(params.STATE.model).forEach(([
-                                                                      i, j
-                                                                    ]) => {
-      const kp1 = keypoints[i];
-      const kp2 = keypoints[j];
-
-      // If score is null, just show the keypoint.
-      const score1 = kp1.score != null ? kp1.score : 1;
-      const score2 = kp2.score != null ? kp2.score : 1;
-      const scoreThreshold = params.STATE.modelConfig.scoreThreshold || 0;
-
-      if (score1 >= scoreThreshold && score2 >= scoreThreshold) {
-        this.ctx.beginPath();
-        this.ctx.moveTo(kp1.x, kp1.y);
-        this.ctx.lineTo(kp2.x, kp2.y);
-        this.ctx.stroke();
-      }
-    });
-  }
-
-  drawKeypoints3D(keypoints) {
-    const scoreThreshold = params.STATE.modelConfig.scoreThreshold || 0;
-    const pointsData =
-        keypoints.map(keypoint => ([-keypoint.x, -keypoint.y, -keypoint.z]));
-
-    const dataset =
-        new scatter.ScatterGL.Dataset([...pointsData, ...ANCHOR_POINTS]);
-
-    const keypointInd =
-        posedetection.util.getKeypointIndexBySide(params.STATE.model);
-    this.scatterGL.setPointColorer((i) => {
-      if (keypoints[i] == null || keypoints[i].score < scoreThreshold) {
-        // hide anchor points and low-confident points.
-        return '#ffffff';
-      }
-      if (i === 0) {
-        return '#ff0000' /* Red */;
-      }
-      if (keypointInd.left.indexOf(i) > -1) {
-        return '#00ff00' /* Green */;
-      }
-      if (keypointInd.right.indexOf(i) > -1) {
-        return '#ffa500' /* Orange */;
-      }
-    });
-
-    if (!this.scatterGLHasInitialized) {
-      this.scatterGL.render(dataset);
-    } else {
-      this.scatterGL.updateDataset(dataset);
-    }
-    const connections = posedetection.util.getAdjacentPairs(params.STATE.model);
-    const sequences = connections.map(pair => ({indices: pair}));
-    this.scatterGL.setSequences(sequences);
-    this.scatterGLHasInitialized = true;
   }
 }
