+++ pose-detection/demos/live_video/src/renderer_webgpu.js
@@ -83,7 +83,8 @@ export class RendererWebGPU {
     return buffer;
   }
 
-  draw(video, tensors, canvasInfo, scoreThreshold) {
+  draw(rendererParams) {
+    const [video, tensors, canvasInfo, scoreThreshold] = rendererParams;
     this.canvasInfo = canvasInfo;
     this.scoreThreshold = scoreThreshold;
     const videoCommanderBuffer = this.drawTexture(video);
@@ -117,7 +118,8 @@ fn main(
   var<function> vertexOutput: VertexOutput;
   let rawY = (keypoints[VertexIndex].x + uniforms.offsetY) * uniforms.scaleY / uniforms.height;
   let rawX  = (keypoints[VertexIndex].y + uniforms.offsetX) * uniforms.scaleX / uniforms.width;
-  var x = rawX * 2.0 - 1.0;
+  // Flip horizontally.
+  var x = 1.0 - rawX * 2.0;
   var y = 1.0 - rawY * 2.0;
   vertexOutput.score = scores[VertexIndex];
   vertexOutput.pos = vec4<f32>(x, y, 1.0, 1.0);
@@ -279,8 +281,11 @@ fn main(@location(0) score: f32) -> @location(0) vec4<f32> {
 
 @fragment fn main(@builtin(position) FragCoord : vec4<f32>)
                          -> @location(0) vec4<f32> {
-    return textureSampleBaseClampToEdge(t, s, FragCoord.xy / vec2<f32>(${
-        this.canvasInfo[4]}, ${this.canvasInfo[5]}));
+    var coord = FragCoord.xy / vec2<f32>(${this.canvasInfo[4]}, ${
+        this.canvasInfo[5]});
+    // Flip horizontally.
+    coord.x = 1.0 - coord.x;
+    return textureSampleBaseClampToEdge(t, s, coord);
 }
       `;
     return [vertexShaderCode, fragmentShaderCode];
